requirements.txt
langgraph>=0.0.31
langgraph-checkpoint-sqlite
langchain>=0.1.13
langchain-google-genai>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
docker>=7.0.0
python-dotenv>=1.0.1
streamlit>=1.30.0
requests>=2.31.0
